Refactor terminal
